Code was initially from apollo client demo, and then it was modified to fit the suggested code structure. Needs additional work
