+++ src/components/__tests__/query-result.js
@@ -1,24 +1,24 @@
-import React from 'react';
-import { render, cleanup } from '../../utils/test-utils';
-import QueryResult from '../query-result';
+import React from "react";
+import { render, cleanup } from "../../utils/test-utils";
+import QueryResult from "../query-result/QueryResult";
 
-describe('Query Result', () => {
+describe("Query Result", () => {
   // automatically unmount and cleanup DOM after the test is finished.
   afterEach(cleanup);
 
-  it('renders loading state', async () => {
+  it("renders loading state", async () => {
     const { getByTestId } = render(<QueryResult loading={true} />);
     getByTestId(/spinner/i);
   });
 
-  it('renders No Data message', async () => {
+  it("renders No Data message", async () => {
     // passing no error and no data
     const { getByText } = render(<QueryResult loading={false} />);
     getByText(/nothing to show/i);
   });
 
-  it('renders Error', async () => {
-    const { getByText } = render(<QueryResult error={new Error('you lose')} />);
+  it("renders Error", async () => {
+    const { getByText } = render(<QueryResult error={new Error("you lose")} />);
     getByText(/you lose/i);
   });
 });
