+++ src/components/content-section/index.js
@@ -1,0 +1,1 @@
+export { default as ContentSection } from "./content-section";
