+++ src/components/header/header.js
@@ -1,0 +1,31 @@
+import React from "react";
+import logo from "../assets/space_cat_logo.png";
+
+/**
+ * Header renders the top navigation
+ * for this particular tutorial level, it only holds the home button
+ */
+const Header = ({ children }) => {
+  return (
+    <HeaderBar>
+      <Container>
+        <HomeButtonContainer>
+          <HomeLink to="/">
+            <HomeButton>
+              <LogoContainer>
+                <Logo src={logo} />
+              </LogoContainer>
+              <Title>
+                <h3>Tool Library</h3>
+                <div>Share tools between neighbors and friends</div>
+              </Title>
+            </HomeButton>
+          </HomeLink>
+        </HomeButtonContainer>
+        {children}
+      </Container>
+    </HeaderBar>
+  );
+};
+
+export default Header;
