+++ src/components/header/styles.js
@@ -1,37 +1,6 @@
-import React from "react";
 import { colors, widths } from "../styles";
 import styled from "@emotion/styled";
 import { Link } from "@reach/router";
-import logo from "../assets/space_cat_logo.png";
-
-/**
- * Header renders the top navigation
- * for this particular tutorial level, it only holds the home button
- */
-const Header = ({ children }) => {
-  return (
-    <HeaderBar>
-      <Container>
-        <HomeButtonContainer>
-          <HomeLink to="/">
-            <HomeButton>
-              <LogoContainer>
-                <Logo src={logo} />
-              </LogoContainer>
-              <Title>
-                <h3>Tool Library</h3>
-                <div>Share tools between neighbors and friends</div>
-              </Title>
-            </HomeButton>
-          </HomeLink>
-        </HomeButtonContainer>
-        {children}
-      </Container>
-    </HeaderBar>
-  );
-};
-
-export default Header;
 
 /** Header styled components */
 const HeaderBar = styled.div({
