+++ src/components/layout/index.js
@@ -1,0 +1,1 @@
+export { default as Layout } from "./layout";
