+++ src/components/layout/layout.js
@@ -1,0 +1,20 @@
+import React from "react";
+import { Header, Footer } from "../components";
+
+/**
+ * Layout renders the full page content:
+ * with header, Page container and footer
+ */
+const Layout = ({ fullWidth, children, grid }) => {
+  return (
+    <>
+      <Header />
+      <PageContainer fullWidth={fullWidth} grid={grid}>
+        {children}
+      </PageContainer>
+      <Footer />
+    </>
+  );
+};
+
+export default Layout;
