+++ src/components/layout/styles.js
@@ -1,0 +1,16 @@
+import styled from "@emotion/styled";
+import { widths, unit } from "../../styles";
+
+/** Layout styled components */
+const PageContainer = styled.div((props) => ({
+  display: "flex",
+  justifyContent: props.grid ? "center" : "top",
+  flexDirection: props.grid ? "row" : "column",
+  flexWrap: "wrap",
+  alignSelf: "center",
+  flexGrow: 1,
+  maxWidth: props.fullWidth ? null : `${widths.regularPageWidth}px`,
+  width: "100%",
+  padding: props.fullWidth ? 0 : unit * 2,
+  paddingBottom: unit * 5,
+}));
