+++ src/components/md-content/index.js
@@ -1,0 +1,1 @@
+export { default as MdContent } from "./md-content";
