+++ src/components/md-content/md-content.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+/**
+ * Markdown component is a simple style wrapper for markdown content used across our app
+ */
+const MarkDown = ({ content }) => {
+  return <StyledMarkdown children={content} />;
+};
+
+export default MarkDown;
