+++ src/components/md-content/styles.js
@@ -1,0 +1,30 @@
+import styled from "@emotion/styled";
+import { colors } from "../../styles";
+import ReactMarkdown from "react-markdown";
+
+/** Markdown styled components */
+const StyledMarkdown = styled(ReactMarkdown)({
+  color: colors.grey.darker,
+
+  h1: {
+    fontSize: "1.7em",
+  },
+  h2: {
+    fontSize: "1.4em",
+  },
+  h3: {
+    fontSize: "1.2em",
+  },
+  a: {
+    color: colors.pink.base,
+  },
+  pre: {
+    padding: 20,
+    borderRadius: 4,
+    border: `solid 1px ${colors.silver.dark}`,
+    backgroundColor: colors.silver.base,
+    code: {
+      fontSize: "0.9em",
+    },
+  },
+});
