+++ src/components/query-result/index.js
@@ -1,0 +1,1 @@
+export { default as QueryResult } from "./query-result";
