+++ src/components/query-result/query-result.js
